noteBackend/requirements.txt
asgiref==3.9.1
click==8.3.0
colorama==0.4.6
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
django-rest-framework==0.1.0
django-restframework==0.0.1
djangorestframework==3.16.1
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
pi==0.1.2
pillow==11.3.0
psycopg2==2.9.11
PyJWT==1.7.1
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2
uvicorn==0.38.0
